Introduce AdminDashboardTabController
